fedimint_ln_client
Module pay

Modules

lightningpay 🔒

lightningpay module is needed to suppress the deprecation warning on the enum declaration. Suppressing the deprecation warning on the enum declaration is not enough, since the derive statement causes it to be ignored for some reason, so instead the enum declaration is wrapped in its own module.

Structs

LightningPayCommon

LightningPayCreatedOutgoingLnContract

LightningPayFunded

LightningPayRefund

LightningPayRefundable

LightningPayStateMachine

PayInvoicePayload

Enums

GatewayPayError

LightningPayStates

State machine that requests the lightning gateway to pay an invoice on behalf of a federation client.

PaymentData

Data needed to pay an invoice, may be the whole invoice or only the required parts of it.

Constants

RETRY_DELAY 🔒

Functions

await_contract_cancelled 🔒

Waits for a contract with contract_id to be cancelled by the gateway.

await_contract_timeout 🔒

Waits until a specific block height at which the contract will be able to be reclaimed.

try_refund_outgoing_contract 🔒

Claims a refund for an expired or cancelled outgoing contract