fedimint_ln_client::pay
Module lightningpay

lightningpay module is needed to suppress the deprecation warning on the enum declaration. Suppressing the deprecation warning on the enum declaration is not enough, since the derive statement causes it to be ignored for some reason, so instead the enum declaration is wrapped in its own module.

Enums

LightningPayStates

State machine that requests the lightning gateway to pay an invoice on behalf of a federation client.