Crate fedimint_ln_client

Modules

api

db

incoming

Incoming State Machine

pay

receive

Structs

LightningClientInit

LightningClientModule

Client side lightning module

LightningOperationMeta

LightningOperationMetaPay

OutgoingLightningPayment

Enums

InternalPayState

The high-level state of an pay operation internal to the federation, started with LightningClientModule::pay_bolt11_invoice.

LightningChildKeys

LightningClientStateMachines

LightningOperationMetaVariant

LnPayState

The high-level state of a pay operation over lightning, started with LightningClientModule::pay_bolt11_invoice.

LnReceiveState

The high-level state of a reissue operation started with LightningClientModule::create_bolt11_invoice.

PayError 🔒

PayType

ReceivingKey

Where to receive the payment to, either to ourselves or to another user

Constants

DEFAULT_INVOICE_EXPIRY_TIME 🔒

META_OVERRIDE_CACHE_DURATION 🔒

OUTGOING_LN_CONTRACT_TIMELOCK 🔒

Number of blocks until outgoing lightning contracts times out and user client can get refund

Functions

create_incoming_contract_output

fetch_and_validate_offer 🔒

invoice_has_internal_payment_markers 🔒

invoice_routes_back_to_federation 🔒

set_payment_result 🔒

tweak_user_key

Tweak a user key with an index, this is used to generate a new key for each invoice. This is done to not be able to link invoices to the same user.

tweak_user_secret_key 🔒

Tweak a secret key with an index, this is used to claim an unspent incoming contract.