Module fedimint_ln_client::incoming

Incoming State Machine

This shared state machine is used by clients that want to pay other clients within the federation

It’s applied in two places:

fedimint-ln-client for internal payments without involving the gateway

gateway for receiving payments into the federation

Structs

AwaitingPreimageDecryption

DecryptingPreimageState

FundingOfferState

IncomingSmCommon

IncomingStateMachine

PreimageState

RefundSuccessState

Enums

IncomingSmError

IncomingSmStates

State machine that executes a transaction between two users within a federation. This creates and funds an incoming contract based on an existing offer within the federation.