Trait fedimint_ln_client::LightningClientExt

pub trait LightningClientExt {
    // Required methods
    fn select_active_gateway<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<LightningGateway>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_active_gateway<'life0, 'life1, 'async_trait>(
        &'life0 self,
        gateway_id: &'life1 PublicKey
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn fetch_registered_gateways<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<LightningGateway>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn pay_bolt11_invoice<'life0, 'async_trait>(
        &'life0 self,
        invoice: Invoice
    ) -> Pin<Box<dyn Future<Output = Result<(PayType, ContractId)>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn subscribe_internal_pay<'life0, 'async_trait>(
        &'life0 self,
        operation_id: OperationId
    ) -> Pin<Box<dyn Future<Output = Result<UpdateStreamOrOutcome<InternalPayState>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn subscribe_ln_pay<'life0, 'async_trait>(
        &'life0 self,
        operation_id: OperationId
    ) -> Pin<Box<dyn Future<Output = Result<UpdateStreamOrOutcome<LnPayState>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn create_bolt11_invoice<'life0, 'async_trait>(
        &'life0 self,
        amount: Amount,
        description: String,
        expiry_time: Option<u64>
    ) -> Pin<Box<dyn Future<Output = Result<(OperationId, Invoice)>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn subscribe_ln_receive<'life0, 'async_trait>(
        &'life0 self,
        operation_id: OperationId
    ) -> Pin<Box<dyn Future<Output = Result<UpdateStreamOrOutcome<LnReceiveState>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Required Methods

fn select_active_gateway<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<LightningGateway>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

The set active gateway, or a random one if none has been set

fn set_active_gateway<'life0, 'life1, 'async_trait>( &'life0 self, gateway_id: &'life1 PublicKey ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Sets the gateway to be used by all other operations

fn fetch_registered_gateways<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<LightningGateway>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Gateways actively registered with the fed

fn pay_bolt11_invoice<'life0, 'async_trait>( &'life0 self, invoice: Invoice ) -> Pin<Box<dyn Future<Output = Result<(PayType, ContractId)>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Pays a LN invoice with our available funds

fn subscribe_internal_pay<'life0, 'async_trait>( &'life0 self, operation_id: OperationId ) -> Pin<Box<dyn Future<Output = Result<UpdateStreamOrOutcome<InternalPayState>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn subscribe_ln_pay<'life0, 'async_trait>( &'life0 self, operation_id: OperationId ) -> Pin<Box<dyn Future<Output = Result<UpdateStreamOrOutcome<LnPayState>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn create_bolt11_invoice<'life0, 'async_trait>( &'life0 self, amount: Amount, description: String, expiry_time: Option<u64> ) -> Pin<Box<dyn Future<Output = Result<(OperationId, Invoice)>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Receive over LN with a new invoice

fn subscribe_ln_receive<'life0, 'async_trait>( &'life0 self, operation_id: OperationId ) -> Pin<Box<dyn Future<Output = Result<UpdateStreamOrOutcome<LnReceiveState>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,